# internal/docs/swagger.yaml
basePath: /
definitions:
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.pagedTorrents:
    properties:
      items:
        items:
          $ref: '#/definitions/seeder.Status'
        type: array
      total:
        type: integer
    type: object
...
  api.speedUpdate:
    properties:
      deleteOnCap:
        type: boolean
      maxRatio:
        type: number
      maxSpeed:
        type: string
    type: object
  api.uploadResult:
    properties:
      id:
        type: string
      name:
        type: string
      notice:
        description: set when tracker URLs were truncated
        type: string
    type: object
...
  seeder.StatsPoint:
    properties:
      l:
        description: total leechers across all trackers
        type: integer
      r:
        description: simulated upload rate, bytes/sec
        type: integer
      t:
        description: unix milliseconds
        type: integer
    type: object
  seeder.Status:
    properties:
      addedAt:
        description: unix milliseconds
        type: integer
      capped:
        description: true when the ratio cap has been reached
        type: boolean
      deleteOnCap:
        description: remove the torrent automatically when capped
        type: boolean
      id:
        description: info hash, hex
        type: string
      location:
        description: .torrent path, relative to the watched root
        type: string
      maxRateBytesPerSec:
        description: configured ceiling
        type: integer
      maxRatio:
        description: upload cap as a multiple of torrent size (0 = unlimited)
        type: number
      name:
        description: torrent display name
        type: string
      rateBytesPerSec:
        type: integer
      sizeBytes:
        type: integer
      trackers:
        items:
          $ref: '#/definitions/seeder.TrackerStatus'
        type: array
      uploadedBytes:
        type: integer
    type: object
  seeder.TrackerStatus:
    properties:
      intervalSec:
        type: integer
      leechers:
        type: integer
      minIntervalSec:
        type: integer
      nextAnnounceAt:
        description: unix ms, 0 if unknown
        type: integer
      seeders:
        type: integer
      status:
        description: '"ok" | "failing" | "pending"'
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: Spoof BitTorrent tracker announces to boost your ratio.
  title: emission API
  version: "1.0"
paths:
...
  /api/torrents:
    get:
      parameters:
      - description: Page size (0 = all, default 10)
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      - description: Case-insensitive name filter
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.pagedTorrents'
      summary: List torrents
      tags:
      - torrents
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Torrent file (.torrent)
        in: formData
        name: file
        required: true
        type: file
      - description: Maximum upload rate (e.g. 1M)
        in: formData
        name: max-speed
        type: string
      - description: Stop uploading at N × torrent size (0 = unlimited)
        in: formData
        name: max-ratio
        type: number
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.uploadResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Upload a .torrent file
      tags:
      - torrents
  /api/torrents/{id}:
    delete:
      parameters:
      - description: Torrent info hash (40-char hex)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Remove a torrent
      tags:
      - torrents
    patch:
      consumes:
      - application/json
      parameters:
      - description: Torrent info hash (40-char hex)
        in: path
        name: id
        required: true
        type: string
      - description: Speed settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.speedUpdate'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Update torrent upload speed
      tags:
      - torrents
  /api/torrents/{id}/stats:
    get:
      parameters:
      - description: Torrent info hash (40-char hex)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/seeder.StatsPoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get torrent stats history
      tags:
      - torrents
...
swagger: "2.0"